refactor: Split window.rs into focused modules
Refactored the 589-line window.rs file into a modular directory structure:

- types.rs: Window function type definitions (89 lines)
- detection.rs: Window function detection helpers (38 lines)
- collection.rs: Window function collection logic (67 lines)
- evaluation.rs: Core evaluation logic (180 lines)
- order_by.rs: ORDER BY window function support (163 lines)
- mod.rs: Public API and orchestration (90 lines)

All files are under 200 lines. Public API preserved (no breaking changes).
All 56 window-related tests pass. Full executor test suite passes (228 tests).

Closes #431
